Move WaitFor into new utils.go and switch timeout and interval to time.Duration.
