+++ acme/dns_challenge.go
@@ -8,7 +8,6 @@ import (
 	"log"
 	"net"
 	"strings"
-	"time"
 
 	"github.com/miekg/dns"
 	"golang.org/x/net/publicsuffix"
@@ -255,27 +254,4 @@ func UnFqdn(name string) string {
 		return name[:n-1]
 	}
 	return name
-}
-
-// WaitFor polls the given function 'f', once every 'interval' seconds, up to 'timeout' seconds.
-func WaitFor(timeout, interval int, f func() (bool, error)) error {
-	var lastErr string
-	timeup := time.After(time.Duration(timeout) * time.Second)
-	for {
-		select {
-		case <-timeup:
-			return fmt.Errorf("Time limit exceeded. Last error: %s", lastErr)
-		default:
-		}
-
-		stop, err := f()
-		if stop {
-			return nil
-		}
-		if err != nil {
-			lastErr = err.Error()
-		}
-
-		time.Sleep(time.Duration(interval) * time.Second)
-	}
 }
